+++ src/main/java/dev/willyelton/crystal_tools/compat/ae2/AppliedEnergisticsCompatability.java
@@ -1,0 +1,5 @@
+package dev.willyelton.crystal_tools.compat.ae2;
+
+public class AppliedEnergisticsCompatability {
+
+}